Fixes for Mac/Clang.

+++ Moco/Moco/MocoCasADiSolver/MocoCasOCProblem.h
@@ -518,7 +518,7 @@ private:
         if (getNumAuxiliaryResidualEquations()) {
             const auto& implicitRefs =
                     mocoProblemRep->getImplicitComponentReferencePtrs();
-            for (int i = 0; i < implicitRefs.size(); ++i) {
+            for (int i = 0; i < (int)implicitRefs.size(); ++i) {
                 const auto& comp = implicitRefs[i].second.getRef();
                 comp.setDiscreteVariableValue(simtkStateDisabledConstraints,
                         implicitRefs[i].first,
@@ -637,7 +637,7 @@ private:
             const auto& residualOutputs =
                     mocoProblemRep.getImplicitResidualReferencePtrs();
             SimTK::Vector auxResiduals((int)residualOutputs.size(), 0.0);
-            for (int i = 0; i < residualOutputs.size(); ++i) {
+            for (int i = 0; i < (int)residualOutputs.size(); ++i) {
                 auxResiduals[i] = residualOutputs[i]->getValue(state);
             }
             std::copy_n(auxResiduals.getContiguousScalarData(),
